requirements.txt
numpy
torch==1.8.1
torchvision==0.9.1
faiss-gpu==1.6.5
pillow
pandas
tqdm
tensorboard
matplotlib
pytorch-metric-learning
timm
scikit-learn
hydra-core==1.1
hydra_colorlog==1.1.0
ipython
jupyter
ipdb
autopep8
flake8
pylint
isort
SciencePlots
